project now works under Windows with MSVC, too
Former-commit-id: d2355f361aa8b6c08e45efc9b8f673fdc61c69bd

+++ src/SAMParser.cc
@@ -36,7 +36,7 @@ void SAMParser::readSAMHeader(void)
             if (foundHeader == false) {
                 throwErrorException("SAM header is missing");
             }
-            size_t offset = -strlen(curr.line);
+            int offset = -((int)strlen(curr.line));
             ifs.seekg(offset, std::ios::cur);
             break;
         }
